+++ tests/test_training.py
@@ -2,12 +2,15 @@ import torch
 from transformers import AdamW
 
 from catwalk import MODELS, TASKS
-from catwalk.steps import FinetuneStep, PredictStep, CalculateMetricsStep
+from catwalk.steps import CalculateMetricsStep, FinetuneStep, PredictStep
+
+from .util import suite_D
 
 
+@suite_D
 def test_training():
-    model = MODELS['rc::tiny-gpt2']
-    task = TASKS['piqa']
+    model = MODELS["rc::tiny-gpt2"]
+    task = TASKS["piqa"]
     instances = task.get_split("train")[:16]
     predictions_before = list(model.predict(task, instances))
     metrics_before = model.calculate_metrics(task, predictions_before)
@@ -32,42 +35,55 @@ def test_training():
 
     predictions_after = list(model.predict(task, instances))
     metrics_after = model.calculate_metrics(task, list(predictions_after))
-    for prediction_before, prediction_after in zip(predictions_before, predictions_after):
-        assert not torch.allclose(prediction_before["acc"][0], prediction_after["acc"][0])
+    for prediction_before, prediction_after in zip(
+        predictions_before, predictions_after
+    ):
+        assert not torch.allclose(
+            prediction_before["acc"][0], prediction_after["acc"][0]
+        )
     assert metrics_before != metrics_after
 
 
+@suite_D
 def test_training_step_gpt():
     finetune_step = FinetuneStep(
-        model='rc::tiny-gpt2',
+        model="rc::tiny-gpt2",
         tasks=["piqa", "sst"],
         train_steps=10,
         validation_steps=10,
     )
     predict_step = PredictStep(model=finetune_step, task="piqa", limit=10)
-    metrics_step = CalculateMetricsStep(model=finetune_step, task="piqa", predictions=predict_step)
+    metrics_step = CalculateMetricsStep(
+        model=finetune_step, task="piqa", predictions=predict_step
+    )
     metrics_step.result()
 
 
+@suite_D
 def test_training_step_t5():
     finetune_step = FinetuneStep(
-        model='rc::t5-very-small-random',
+        model="rc::t5-very-small-random",
         tasks=["rte", "boolq"],
         train_steps=10,
         validation_steps=10,
     )
     predict_step = PredictStep(model=finetune_step, task="rte", limit=10)
-    metrics_step = CalculateMetricsStep(model=finetune_step, task="rte", predictions=predict_step)
+    metrics_step = CalculateMetricsStep(
+        model=finetune_step, task="rte", predictions=predict_step
+    )
     metrics_step.result()
 
 
+@suite_D
 def test_training_step_hf():
     finetune_step = FinetuneStep(
-        model='tiny-bert',
+        model="tiny-bert",
         tasks=["piqa"],
         train_steps=10,
         validation_steps=10,
     )
     predict_step = PredictStep(model=finetune_step, task="piqa", limit=10)
-    metrics_step = CalculateMetricsStep(model=finetune_step, task="piqa", predictions=predict_step)
+    metrics_step = CalculateMetricsStep(
+        model=finetune_step, task="piqa", predictions=predict_step
+    )
     metrics_step.result()