+++ tests/util.py
@@ -1,0 +1,17 @@
+import pytest
+
+
+def suite_A(test_method):
+    return pytest.mark.suite_A(test_method)
+
+
+def suite_B(test_method):
+    return pytest.mark.suite_B(test_method)
+
+
+def suite_C(test_method):
+    return pytest.mark.suite_C(test_method)
+
+
+def suite_D(test_method):
+    return pytest.mark.suite_D(test_method)